pyproject.toml
[tool.poetry]
name = "banking-api-leimtb"
version = "0.1.0"
description = "sample Bank API"
authors = ["Pavel Nikulin <p.nikulin@innopolis.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "0.11.3"
fastapi = "0.89.0"
python-multipart = "0.0.5"
email-validator = "1.0.5"
requests = "2.23.0"
passlib = {extras = ["bcrypt"], version = "1.7.2"}
pydantic = {extras = ["dotenv"], version = "1.10.4"}
gunicorn = "20.0.4"
psycopg2-binary = "^2.8.5"
alembic = "1.4.2"
sqlalchemy = "1.4.41"
python-jose = {extras = ["cryptography"], version = "3.1.0"}
fastapi-pagination = {extras = ["all"], version = "^0.11.2"}

[tool.poetry.dev-dependencies]
mypy = "0.770"
black = "19.10b0"
isort = "4.3.21"
autoflake = "1.3.1"
flake8 = "3.7.9"
pytest = "5.4.1"
sqlalchemy-stubs = "0.3"
pytest-cov = "2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

